Simple Calculator
This project utilizes and showcases some basic features in Tkinter. The project consist of but not limited to Buttons, Entry Feilds, grid manipulation as well as labels and functions.
This will be open-source so anyone could use these as an example or if anyone would like to add feedback. Happy coding to all!